Stage 2
Word search at this point is fully functional. We display to the user what words they need to look for. Currently, for development purposes, the word is highlighted in light gray. Any changes we make past this will be bonus features. 

We are looking to add a function where users can select the boxes and we will let them know if they found the word or not.